added type conversion

+++ 1_1/High-Level Programming I: The C Programming Language/knk_c/14 The Preprocessor/Exercises/ex02.c
@@ -2,7 +2,7 @@
 
 #include <stdio.h>
 
-#define NELEMS(a) sizeof(a) / sizeof(a[0])
+#define NELEMS(a) ((int) sizeof(a) / sizeof(a[0]))
 
 int main(void) {
     int a[] = {1, 2, 3, 4};
